Added Rating filter, changed application of filters and minor darkmode detail #60
Added the review filters, so that movie can be filtered to have a minimum score.

How filters are applied had to be changed to prevent an issue when loading the movies once a single filter has been removed that was applied and visible under the search bar.

The minor detail for the darkmode that  was added is that the text on the current page was not really showing when the site had darkmode activated. Changed it to white.